fix: calendars list can be lost if interrupted
entering a create, delete, or go to date dialog before the initial
calendars list response was returned would cause the calendars list to
be lost since the model would route the calendars list response to the
dialog instead of the calendar where it would be discarded. this
forcibly routes those messages to the calendar so the list can be saved.

note, events responses are not forcibly routed to the calendar since
exiting dialogs triggers a new events request anyway